Fixed issue of bot quoting itself
When a command was invoking that involved the bot sending multiple messages, like asking for confirmation, the bot quoted its previous message.
Looked bad. Fixed it.
Changed the show_stats method in character.py to return the stat table instead of sending it. and changed some messages to be send using ctx.channel.send instead of ctx.send.
that fixed the issue of the bot quoting itself.